% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tags.R
\docType{package}
\name{FSKA-package}
\alias{FSKA-package}
\alias{FSKA}
\title{Fast Sum-Kernel Approximation}
\description{
Package with a small example of fast approximation of the sum-kernel
problem. See the README.
}
