% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{get_ess}
\alias{get_ess}
\title{Effective Sample size of a mssm Object}
\usage{
get_ess(object)
}
\arguments{
\item{object}{an object of class \code{mssm}.}
}
\value{
An object of class \code{mssmEss} with the effective sample sizes.
}
\description{
Extracts the effective sample size at each time point from a \code{mssm}
object.
}
