% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{logLik.mssm}
\alias{logLik.mssm}
\title{Approximate Log-likelihood for a mssm Object}
\usage{
\method{logLik}{mssm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mssm}.}

\item{...}{un-used.}
}
\value{
A \code{logLik} object. The \code{log_lik_terms} attribute contains
the log-likelihood contributions from each time point.
}
\description{
Function to extract the log-likelihood from a \code{mssm} object.
}
