% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-tags.R
\docType{package}
\name{mssm-package}
\alias{mssm-package}
\title{Multivariate State Space Models}
\description{
This package contains particle filter methods for multivariate observed
outcomes and low dimensional state vectors. The methods are intended to
scale well in the dimension of the observed outcomes. The main function in
the package is the \code{\link{mssm}} function.

The package is still under development and the API and results of the
methods may change.
}
