% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm-pf}
\alias{mssm-pf}
\title{Particle Filter Function for Multivariate State Space Model}
\arguments{
\item{cfix}{coefficient for fixed effects.}

\item{disp}{additional parameters for the family (e.g., a dispersion
parameter).}

\item{F.}{matrix in the transition density of the state vector.}

\item{Q}{covariance matrix in the transition density of the state vector.}

\item{Q0}{optional covariance matrix at the first time point. Default is
the covariance matrix in the time invariant distribution.}

\item{mu0}{optional mean at the first time point. Default is
the zero vector.}

\item{trace}{integer controlling whether information should be printed
during particle filtering. Zero yields no information.}

\item{seed}{integer to pass to \code{\link{set.seed}}. The seed is not set
if the argument is \code{NULL}.}

\item{what, N_part}{as as in \code{\link{mssm_control}}.}
}
\value{
An object of class \code{mssm} with the following elements
\item{pf_output}{A list with an element for each time period. Each element
is a list with
\code{particles}: the sampled particles,
\code{stats}: additional object that is requested to be computed with
each particle,
\code{ws:} unnormalized log particle weights for the filtering distribution,
and
\code{ws_normalized:} normalized log particle weights for the filtering
distribution.}
\item{remaining elements}{same as return by \code{\link{mssm}}.}

If gradient approximation is requested then the first elements of
\code{stats} are w.r.t. the fixed coefficients, the next elements are
w.r.t. the matrix in the map from the previous state vector to the mean
of the next, and the last element is w.r.t. the covariance matrix.
Only the lower triangular matrix is kept for the covariance
matrix. See the examples in the README at
\url{https://github.com/boennecd/mssm}.

If an approximation of the observed information matrix is requested then
each of the two gradient parts for the state equation and observational
equation are followed by the corresponding elements
of the approximation of the observed information matrix. That is, if we
have \eqn{p} fixed coefficient then the first \eqn{p} elements is the
gradient elements and the next \eqn{p^2} elements is the first block diagonal
matrix of the observed information matrix. Next, if we have a \eqn{q}
dimensional state vector then the first \eqn{q^2 + q(q + 1) /2} elements
are the gradient w.r.t. the parameters and the remaining elements is the
last block matrix of the observed information matrix.
}
\description{
Function returned from \code{\link{mssm}} which can be used to perform
particle filtering given values for the parameters in the model.
}
