% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm-smoother}
\alias{mssm-smoother}
\title{Computes Smoothed Particle Weights for Multivariate State Space
Model}
\arguments{
\item{object}{an object of class \code{mssm} from \link{mssm-pf}.}
}
\value{
Same as \link{mssm-pf} but where the \code{pf_output}'s list elements
has an additional element called \code{ws_normalized_smooth}. This
contains the normalized log smoothing weights.
}
\description{
Computes smoothed weights using the backward smoothing formula for a
\code{mssm} object. The k-d dual tree approximation is also used if it used
for the \code{mssm} object.
}
\seealso{
\code{\link{mssm}}.
}
