% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mssm.asJavaMatrix}
\alias{mssm.asJavaMatrix}
\title{Transform a matrix into a Java Matrix}
\usage{
mssm.asJavaMatrix(matrix)
}
\arguments{
\item{matrix}{a matrix to be transformed into a Java Matrix.}
}
\value{
A Java two-dimensionnal array of double (called Java Matrix).
}
\description{
Produces a Java two-dimensionnal array of double (a Java Matrix) from
a R matrix.
Java Matrices are the standard type used for subsequent analysis with
the \link{mssm} package.
}
\section{Maintainer}{
 Vincent Branders <vincent.branders@uclouvain.be>.
}

\seealso{
\code{\link[=mssm.fromJavaMatrix]{mssm.fromJavaMatrix()}} to transform a Java Matrix into a
matrix and \code{\link[=mssm.loadMatrix]{mssm.loadMatrix()}} to load a Java Matrix from a file.
See also package \link{mssm}.
}
