% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm.getHeuristicReordering}
\alias{mssm.getHeuristicReordering}
\title{Heuristic ordering of the columns}
\usage{
mssm.getHeuristicReordering(x)
}
\arguments{
\item{x}{a matrix or a Java two-dimensionnal array of double (\code{\link[=mssm.asJavaMatrix]{mssm.asJavaMatrix()}}).}
}
\value{
A vector of the columns indices to be investigated first (starting from zero).
}
\description{
Returns an ordering of the column indices to improve the search of
\code{\link[=mssm.search.cpgc]{mssm.search.cpgc()}}.
Columns are sorted on the descending sum of their positive entries.
This sum is computed as the sum of the entries (in the column) that
are positive.
The purpose of this function is to provide one (among many possible)
heuristic ordering of the columns to be used in the subsequent search.
}
\section{Maintainer}{
 Vincent Branders <vincent.branders@uclouvain.be>.
}

\seealso{
\code{\link[=mssm.asJavaMatrix]{mssm.asJavaMatrix()}} and \code{\link[=mssm.search.cpgc]{mssm.search.cpgc()}} for the
transformation of a matrix to a Java Matrix and the search of a submatrix
of maximal sum. See also package \link{mssm}.
}
