% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mssm.isValid}
\alias{mssm.isValid}
\title{Ensures type of objects}
\usage{
mssm.isValid(object, class, silent = F)
}
\arguments{
\item{object}{a Java object to be validated.}

\item{class}{the type that should be matched by \code{object}.}

\item{silent}{Boolean preventing (\code{TRUE}) error message is not valid.}
}
\value{
A boolean indicating if the object is of expected type.
}
\description{
Ensures that object \code{object} is of type \code{class} and return a boolean. If type is not valid, it generates a warning.
}
\section{Maintainer}{
 Vincent Branders <vincent.branders@uclouvain.be>.
}

\seealso{
\code{\link{.jinstanceof}}.
}
