% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mssm.jInst}
\alias{mssm.jInst}
\title{Instanciation of Java objects}
\usage{
mssm.jInst(name)
}
\arguments{
\item{name}{a string corresponding to the name of the class as
indicated by \code{\link[=mssm.listClasses]{mssm.listClasses()}}.}
}
\value{
A Java instance of the named class.
}
\description{
This function retrieves a Java instance of a named class.
The class names and locations are listed by \code{\link[=mssm.listClasses]{mssm.listClasses()}}.
}
\seealso{
\code{\link{mssm.listClasses}} for a list of possible class
names. See also package \link{mssm}.
}
