% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mssm.listClasses}
\alias{mssm.listClasses}
\title{Lists Java classes of the \link{mssm} package}
\usage{
mssm.listClasses()
}
\value{
It returns a named vactor of the class locations.
}
\description{
Lists all Java classes and their associated location.
}
\section{Maintainer}{
 Vincent Branders <vincent.branders@uclouvain.be>.
}

\seealso{
\code{\link[=mssm.jInst]{mssm.jInst()}} for the instanciation of java Objects.
See also package \link{mssm}.
}
