% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm}
\alias{mssm}
\title{Get Multivariate State Space Model Function}
\usage{
mssm(fixed, family, data, random, weights, offsets, ti,
  control = mssm_control())
}
\arguments{
\item{fixed}{\code{\link{formula}} with outcome variable on the left hand
side and covariates with fixed effects on the right hand side.}

\item{family}{family for the observed outcome given the state variables
and covariates.}

\item{data}{\code{\link{data.frame}} or environment containing the variables
in \code{fixed} and \code{random}.}

\item{random}{\code{\link{formula}} for covariates with a random effect.
Left hand side is ignored.}

\item{weights}{optional prior weights.}

\item{offsets}{optional a priori known component in the linear predictor.}

\item{ti}{integer vector with time indices matching with each observation of
\code{fixed} and \code{random}.}

\item{control}{list with arguments passed to \code{\link{mssm_control}}.}
}
\value{
An object of class \code{mssmFunc} with the following elements
\item{pf_filter}{function to perform particle filtering. See
\link{mssm-pf}.}
\item{terms_fixed}{\code{\link{terms.object}} for the covariates with
fixed effects.}
\item{terms_random}{\code{\link{terms.object}} for the covariates with
random effects.}
\item{y}{vector with outcomes.}
\item{X}{covariates with fixed effects.}
\item{Z}{covariates with random effects.}
\item{ti}{time indices for each observation.}
\item{weights}{prior weights for each observation.}
\item{offsets}{a priori known component in the linear predictor for
each observation.}
\item{call}{the matched call.}
\item{family}{character describing the conditional distribution of the
outcomes.}
}
\description{
Returns an object which can be used to run a particle filter.
}
\seealso{
The README of the package contains examples of how to use this function.
See \url{https://github.com/boennecd/mssm}.
}
