% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mssmEss}
\alias{plot.mssmEss}
\title{Plot Effective Sample Sizes}
\usage{
\method{plot}{mssmEss}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{mssmEss}.}

\item{y}{un-used.}

\item{...}{un-used.}
}
\value{
The plotted x-values, y-values, and maximum possible effective sample
size.
}
\description{
Plots the effective sample sizes.
}
