% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\docType{package}
\name{mssm-package}
\alias{mssm}
\alias{mssm-package}
\title{mssm: Sub-Matrix of Maximal Sum identification, Scala implementation wrapped in R}
\description{
This package provides tools for the identification of submatrix of maximal sum.
    A submatrix is a rectangular, non-necessarily contiguous, subset of rows and of columns of a (larger) matrix.
    The submatrices that are searched for are max-sum submatrices.
    A submatrix is of maximal sum if the sum of its entires (selected from the original matrix) is of maximal sum.
    This package is a wrapper for the Scala implementation originally provided.
}
\details{
You are most likely to use \code{\link[=mssm.search.cpgc]{mssm.search.cpgc()}} for the seach.
You might require \code{\link[=mssm.loadMatrix]{mssm.loadMatrix()}} to load matrices into a Java
two-dimensionnal array of Double.
Alternatively, functons \code{\link[=mssm.asJavaMatrix]{mssm.asJavaMatrix()}} and \code{\link[=mssm.fromJavaMatrix]{mssm.fromJavaMatrix()}}
allows conversion fom R to Java and vice versa.
}
\keyword{internal}
