% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm-pf}
\alias{mssm-pf}
\title{Particle Filter Function for Multivariate State Space Model}
\arguments{
\item{cfix}{coefficient for fixed effects.}

\item{disp}{additional parameters for the family (e.g., a dispersion
parameter).}

\item{F.}{matrix in the transition density of the state vector.}

\item{Q}{covariance matrix in the transition density of the state vector.}

\item{Q0}{optional covariance matrix at the first time point. Default is
the covariance matrix in the time invariant distribution.}

\item{mu0}{optional mean at the first time point. Default is
the zero vector.}

\item{trace}{integer controlling whether information should be printed
during particle filtering. Zero yields no information.}

\item{seed}{integer to pass to \code{\link{set.seed}}. The seed is not set
if the argument is \code{NULL}.}

\item{what, N_part}{same as in \code{\link{mssm_control}}.}
}
\value{
An object of class \code{mssm} with the following elements
\item{pf_output}{A list with an element for each time period. Each element
is a list with
\code{particles}: the sampled particles,
\code{stats}: additional object that is requested to be computed with
each particle,
\code{ws:} unnormalized log particle weights for the filtering distribution,
and
\code{ws_normalized:} normalized log particle weights for the filtering
distribution.}

Remaining elements are the same as returned by \code{\link{mssm}}.

If gradient approximation is requested then the first elements of
\code{stats} are w.r.t. the fixed coefficients, the next elements are
w.r.t. the matrix in the map from the previous state vector to the mean
of the next, and the last element is w.r.t. the covariance matrix.
Only the lower triangular matrix is kept for the covariance
matrix. See the examples in the README at
\url{https://github.com/boennecd/mssm}. There will be an additional
element for the dispersion parameter if the family has a dispersion
parameter.

If an approximation of the observed information matrix is requested then
it's components are given after the gradient elements in the
\code{stats} object.
}
\description{
Function returned from \code{\link{mssm}} which can be used to perform
particle filtering given values for the parameters in the model.
}
\seealso{
\code{\link{mssm}}.
}
