% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{mssm.getUpperBound}
\alias{mssm.getUpperBound}
\title{Computes the upper bound of a matrix}
\usage{
mssm.getUpperBound(javaMatrix)
}
\arguments{
\item{javaMatrix}{a Java two-dimensionnal array of double.}
}
\value{
A numeric value corresponding to the sum of all positive values in
the Java matrix.
}
\description{
Computes the upper bound of a matrix. The upper bound is computed here
as the sum of all positive values.
This can be usefull to determines if the found solution match the upper
bound computed here as this means that the solution is an optimal solution.
}
\seealso{
\code{\link[=mssm.asJavaMatrix]{mssm.asJavaMatrix()}} and \code{\link[=mssm.search.cpgc]{mssm.search.cpgc()}} for the
transformation of a matrix to a Java Matrix and the search of a submatrix
of maximal sum. See also package \link{mssm}.
}
