% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssm.R
\name{plot.mssm}
\alias{plot.mssm}
\title{Plot Predicted State Variables for mssm Object.}
\usage{
\method{plot}{mssm}(x, y, qs = c(0.05, 0.95), do_plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{mssm}.}

\item{y}{un-used.}

\item{qs}{two-dimensional numeric vector with bounds of the prediction
interval.}

\item{do_plot}{\code{TRUE} to create a plot with the mean and quantiles.}

\item{...}{un-used.}
}
\value{
List with means and quantiles.
}
\description{
Plots the predicted mean and pointwise prediction interval of the state
variables for the filtering distribution.
}
